Remove Last ContextAPI Settings
